requirements.txt
# Core dependencies
langchain>=0.1.0
pika>=1.3.2  # For RabbitMQ
PyMuPDF>=1.23.7  # For PDF parsing
sentence-transformers>=2.2.2  # For embeddings
numpy>=1.24.0
pydantic>=2.5.0

# Database
SQLAlchemy>=2.0.0

# LLM
langchain-community>=0.0.13
langchain-ollama>=0.0.1

# Utils
python-dotenv>=1.0.0